backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
chromadb
python-dotenv
pydantic
pypdf
groq
slowapi
